fixing initialization
