+++ watertap/core/membrane_channel0d.py
@@ -25,10 +25,13 @@ from idaes.core import (
     EnergyBalanceType,
     FlowDirection,
 )
+from idaes.core.solvers import get_solver
 from idaes.core.util import scaling as iscale
 from idaes.core.util.misc import add_object_reference
 from idaes.core.util.exceptions import BalanceTypeNotSupportedError
 from idaes.core.base.control_volume0d import ControlVolume0DBlockData
+import idaes.logger as idaeslog
+
 from watertap.core.membrane_channel_base import (
     MembraneChannelMixin,
     PressureChangeType,
@@ -163,6 +166,91 @@ class MembraneChannel0DBlockData(MembraneChannelMixin, ControlVolume0DBlockData)
             raise ConfigurationError(
                 f"Unrecognized pressure_change_type {pressure_change_type}"
             )
+
+    def initialize(
+        self,
+        state_args=None,
+        outlvl=idaeslog.NOTSET,
+        optarg=None,
+        solver=None,
+        hold_state=True,
+        initialize_guess=None,
+    ):
+        """
+        Initialization routine for the membrane channel control volume
+
+        Keyword Arguments:
+            state_args : a dict of arguments to be passed to the property
+                         package(s) to provide an initial state for
+                         initialization (see documentation of the specific
+                         property package) (default = {}).
+            outlvl : sets output log level of initialization routine
+            optarg : solver options dictionary object (default=None, use
+                     default solver options)
+            solver : str indicating which solver to use during
+                     initialization (default = None)
+            hold_state : flag indicating whether the initialization routine
+                     should unfix any state variables fixed during
+                     initialization, **default** - True. **Valid values:**
+                     **True** - states variables are not unfixed, and a dict of
+                     returned containing flags for which states were fixed
+                     during initialization, **False** - state variables are
+                     unfixed after initialization by calling the release_state
+                     method.
+            initialize_guess : a dict of guesses for solvent_recovery, solute_recovery,
+                     and cp_modulus. These guesses offset the initial values
+                     for the retentate, permeate, and membrane interface
+                     state blocks from the inlet feed
+                     (default =
+                     {'deltaP': -1e4,
+                     'solvent_recovery': 0.5,
+                     'solute_recovery': 0.01,
+                     'cp_modulus': 1.1})
+
+        Returns:
+            If hold_states is True, returns a dict containing flags for which
+            states were fixed during initialization.
+        """
+        if optarg is None:
+            optarg = {}
+        # Create solver
+        opt = get_solver(solver, optarg)
+
+        # Get inlet state if not provided
+        init_log = idaeslog.getInitLogger(self.name, outlvl, tag="control_volume")
+        solve_log = idaeslog.getSolveLogger(self.name, outlvl, tag="control_volume")
+
+        state_args = self._get_state_args(initialize_guess, state_args)
+
+        # intialize self.properties
+        source_flags = self.properties_in.initialize(
+            state_args=state_args["feed_side"],
+            outlvl=idaeslog.ERROR,
+            optarg=optarg,
+            solver=solver,
+            hold_state=True,
+        )
+
+        self.properties_out.initialize(
+            state_args=state_args["retentate"],
+            outlvl=outlvl,
+            optarg=optarg,
+            solver=solver,
+        )
+
+        self.properties_interface.initialize(
+            outlvl=outlvl,
+            optarg=optarg,
+            solver=solver,
+            state_args=state_args["interface"],
+        )
+
+        init_log.info("Initialization Complete")
+
+        if hold_state:
+            return source_flags
+        else:
+            self.release_state(source_flags, outlvl)
 
     def calculate_scaling_factors(self):
         super().calculate_scaling_factors()
